Exercise 217
Some of the function use logic that tests for certain states and thus does not
follow the design recipe, which says to design functions according to the data
definitions they work on. This could be changed if at the start of the game the
worm segments all had a direction vector set to one of the legal values
and not to (make-dir 0 0).