Exercises 425, 426, 427, 428
Removed (rest list-of-x) in the recursive calls of quick-sort, since the
function smallers and largers already remove the first element
of the list.
Additionally, one conditional in quick-sort in exercises 427 and 428 has
been removed, as that case is taken care of by a different sorting
function when the length of the list is below the threshold. The case for
the empty list has to remain, since the threshold can be set to 0.